Make the background brightness configurable
Now you can `let g:oxidize_brightness` between `1` and `10` to increase
the background brightness for lower contrast and better visibility in
bright conditions. `0` is classic black, and `4` looks pretty good when
working in the sunlight.

In fact, I like `2` so much that I made it the new default after ten
years of having a black background!